+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#1176](https://github.com/alleslabs/celatone-frontend/pull/1176) Improve adding custom network by url with query param
 - [#1175](https://github.com/alleslabs/celatone-frontend/pull/1175) Fix tooltip bg color
 - [#1172](https://github.com/alleslabs/celatone-frontend/pull/1172) Fix auto popup for module drawer in mobile